requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
sentence-transformers>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
